proxy-function/requirements.txt
flask
requests
functions-framework==3.*
google-cloud-secret-manager